refactor(lockfile): split package source parsing
